+++ nukibridge.js
@@ -1,6 +1,5 @@
 var request = require("request");
 var http = require('http');
-var url = require('url');
 
 global.NUKI_LOCK_ACTION_UNLOCK = "1";
 global.NUKI_LOCK_ACTION_LOCK = "2";
@@ -10,11 +9,11 @@ global.NUKI_LOCK_ACTION_LOCK_N_GO_UNLATCH = "5";
 
 global.NUKI_LOCK_STATE_UNCALIBRATED = 0;
 global.NUKI_LOCK_STATE_LOCKED = 1;
-global.NUKI_LOCK_STATE_UNLOCKED = 2;
-global.NUKI_LOCK_STATE_UNLOCKED_LOCK_N_GO = 3;
-global.NUKI_LOCK_STATE_UNLATCHED = 4;
-global.NUKI_LOCK_STATE_LOCKING = 5;
-global.NUKI_LOCK_STATE_UNLOCKING = 6;
+global.NUKI_LOCK_STATE_UNLOCKING = 2;
+global.NUKI_LOCK_STATE_UNLOCKED = 3;
+global.NUKI_LOCK_STATE_LOCKING = 4;
+global.NUKI_LOCK_STATE_UNLATCHED = 5;
+global.NUKI_LOCK_STATE_UNLOCKED_LOCK_N_GO = 6;
 global.NUKI_LOCK_STATE_UNLATCHING = 7;
 global.NUKI_LOCK_STATE_MOTOR_BLOCKED = 254;
 global.NUKI_LOCK_STATE_UNDEFINED = 255;
@@ -24,7 +23,7 @@ var DEFAULT_REQUEST_TIMEOUT_LOCK_ACTION = 30000;
 var DEFAULT_WEBHOOK_SERVER_PORT = 51827;
 var DEFAULT_CACHE_DIRECTORY = "./.node-persist/storage";
 
-function NukiBridge(log, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory, webHookServerPort) {
+function NukiBridge(log, bridgeUrl, apiToken, requestTimeoutLockState, requestTimeoutLockAction, cacheDirectory, webHookServerIpOrName, webHookServerPort) {
     this.log = log;
     this.log("Initializing Nuki bridge '%s'...", bridgeUrl);
     this.bridgeUrl = bridgeUrl;
@@ -32,6 +31,7 @@ function NukiBridge(log, bridgeUrl, apiToken, requestTimeoutLockState, requestTi
     this.requestTimeoutLockState = requestTimeoutLockState;
     this.requestTimeoutLockAction = requestTimeoutLockAction;
     this.cacheDirectory = cacheDirectory;
+    this.webHookServerIpOrName = webHookServerIpOrName;
     this.webHookServerPort = webHookServerPort;
     if(this.requestTimeoutLockState == null || this.requestTimeoutLockState == "" || this.requestTimeoutLockState < 1) {
         this.requestTimeoutLockState = DEFAULT_REQUEST_TIMEOUT_LOCK_STATE;
@@ -52,51 +52,52 @@ function NukiBridge(log, bridgeUrl, apiToken, requestTimeoutLockState, requestTi
     this.queue = [];
     this.locks = [];
     
-    this._createWebHookServer(this.log, this.webHookServerPort);
+    if(this.webHookServerIpOrName && this.webHookServerIpOrName !== "") {
+        this.webHookUrl = "http://"+this.webHookServerIpOrName+":"+this.webHookServerPort+"/";
+        this._createWebHookServer(this.log, this.webHookServerPort);
+        this._addWebhookToBridge();
+    }
     
     this.log("Initialized Nuki bridge.");
 };
 
 NukiBridge.prototype._createWebHookServer = function _createWebHookServer(log, webHookServerPort) {
     http.createServer((function(request, response) {
-        var theUrl = request.url;
-        var theUrlParts = url.parse(theUrl, true);
-        var theUrlParams = theUrlParts.query;
         var body = [];
         request.on('error', function(err) {
-            console.error("[ERROR Nuki WebHook Server] Reason: %s.", err);
+            log("[ERROR Nuki WebHook Server] Reason: %s.", err);
         }).on('data', function(chunk) {
             body.push(chunk);
         }).on('end', (function() {
             body = Buffer.concat(body).toString();
 
             response.on('error', function(err) {
-                console.error("[ERROR Nuki WebHook Server] Reason: %s.", err);
+                log("[ERROR Nuki WebHook Server] Reason: %s.", err);
             });
-
+            
             response.statusCode = 200;
             response.setHeader('Content-Type', 'application/json');
-
-            if(!theUrlParams.nukiId || !theUrlParams.state || !theUrlParams.batteryCritical) {
+            
+            var json = JSON.parse(body);
+            if(!json.nukiId || !json.state) {
                 response.statusCode = 404;
                 response.setHeader("Content-Type", "text/plain");
                 var errorText = "[ERROR Nuki WebHook Server] No nukiId or state or batteryCritical in request.";
-                console.error(errorText);
+                log(errorText);
                 response.write(errorText);
                 response.end();
             }
             else {
-                var nukiId = theUrlParams.nukiId;
-                var state = theUrlParams.state;
-                var batteryCritical = theUrlParams.batteryCritical===true || theUrlParams.batteryCritical==="true";
+                var nukiId = json.nukiId+"";
+                var state = json.state;
+                var batteryCritical = json.batteryCritical===true || json.batteryCritical==="true";
                 var ignoreDoorsWithDoorLatches = true;
                 var lock = this._getLock(nukiId, ignoreDoorsWithDoorLatches);
                 if(lock == null) {
-                    response.statusCode = 404;
                     response.setHeader("Content-Type", "text/plain");
-                    var errorText = "[ERROR Nuki WebHook Server] No lock found for nukiId '"+nukiId+"'.";
-                    console.error(errorText);
-                    response.write(errorText);
+                    var infoText = "[INFO Nuki WebHook Server] No lock found for nukiId '"+nukiId+"'.";
+                    log(infoText);
+                    response.write(infoText);
                     response.end();
                 }
                 else {            
@@ -106,7 +107,7 @@ NukiBridge.prototype._createWebHookServer = function _createWebHookServer(log, w
                     if(!lock.isDoorLatch()) {
                         var isLocked = lock._isLocked(state);
                         lock._setLockCache(isLocked, batteryCritical);  
-                        console.log("[INFO Nuki WebHook Server] Updated lock state from webhook to isLocked = '%s' (Nuki state '%s' ) for lock '%s' (instance id '%s') with batteryCritical = '%s'.", isLocked, state, lock.id, lock.instanceId, batteryCritical);
+                        log("[INFO Nuki WebHook Server] Updated lock state from webhook to isLocked = '%s' (Nuki state '%s' ) for lock '%s' (instance id '%s') with batteryCritical = '%s'.", isLocked, state, lock.id, lock.instanceId, batteryCritical);
                         lock.webHookCallback(isLocked, batteryCritical);
                     }                      
                     response.write(JSON.stringify(responseBody));
@@ -116,6 +117,69 @@ NukiBridge.prototype._createWebHookServer = function _createWebHookServer(log, w
         }).bind(this));
     }).bind(this)).listen(webHookServerPort);    
     this.log("Started server for webhooks on port '%s'.", webHookServerPort);
+};
+
+NukiBridge.prototype._addWebhookToBridge = function _addWebhookToBridge() {
+    this.log("Adding webhook for plugin to bridge...");
+    var callbackList = (function(err, json) {
+        if(err) {
+            throw new Error("Request for webhooks failed: "+err);
+        }
+        else if(json) {
+            var callbacks = json.callbacks;
+            var webhookExists = false;
+            for (var i = 0; i < callbacks.length; i++) {
+                var callback = callbacks[i];
+                if(callback.url === this.webHookUrl) {
+                    webhookExists = true;
+                    break;
+                }
+            }
+            if(webhookExists) {
+                this.log("Webhook for plugin already exists.", err);
+            }
+            else {
+                this._addCallback();
+            }
+        }
+    }).bind(this);
+    this._getCallbacks(callbackList);
+};
+
+NukiBridge.prototype._getCallbacks = function _getCallbacks(callback) {
+    if(!this.runningRequest) {
+        this._sendRequest(
+            "/callback/list",
+            { token: this.apiToken},
+            this.requestTimeoutLockAction,
+            callback
+        );
+    }
+    else {
+        this._addToQueue({callbackList: callbackList});
+    }
+};
+
+NukiBridge.prototype._addCallback = function _addCallback() {
+    if(!this.runningRequest) {
+        var callback = (function(err, json) {
+            if(err) {
+                throw new Error("Adding webhook failed: "+err);
+            }
+            else {
+                this.log("Webhook for plugin added.", err);
+            }
+        }).bind(this);
+        this._sendRequest(
+            "/callback/add",
+            { token: this.apiToken, url: this.webHookUrl},
+            this.requestTimeoutLockAction,
+            callback
+        );
+    }
+    else {
+        this._addToQueue({callbackAdd: true});
+    }
 };
 
 NukiBridge.prototype.addLock = function addLock(nukiLock) {
@@ -139,7 +203,6 @@ NukiBridge.prototype._getLock = function _getLock(id, ignoreDoorsWithDoorLatches
 NukiBridge.prototype.lockState = function lockState(nukiLock, callback /*(err, json)*/) {
     if(!this.runningRequest) {
         this._sendRequest(
-            nukiLock,
             "/lockState",
             { token: this.apiToken, nukiId: nukiLock.id},
             this.requestTimeoutLockState,
@@ -153,9 +216,8 @@ NukiBridge.prototype.lockState = function lockState(nukiLock, callback /*(err, j
 
 NukiBridge.prototype.lockAction = function lockAction(nukiLock, lockAction, callback /*(err, json)*/) {
     if(!this.runningRequest) {
-        nukiLock.log("Process lock action '%s' for Nuki lock '%s' (instance id '%s') on Nuki bridge '%s'.", lockAction, nukiLock.id, nukiLock.instanceId, this.bridgeUrl);
+        this.log("Process lock action '%s' for Nuki lock '%s' (instance id '%s') on Nuki bridge '%s'.", lockAction, nukiLock.id, nukiLock.instanceId, this.bridgeUrl);
         this._sendRequest(
-            nukiLock,
             "/lockAction",
             { token: this.apiToken, nukiId: nukiLock.id, action: lockAction},
             this.requestTimeoutLockAction,
@@ -167,8 +229,8 @@ NukiBridge.prototype.lockAction = function lockAction(nukiLock, lockAction, call
     }
 };
 
-NukiBridge.prototype._sendRequest = function _sendRequest(nukiLock, entryPoint, queryObject, requestTimeout, callback /*(err, json)*/) {
-    nukiLock.log("Send request to Nuki bridge '%s' on '%s' with '%s'.", this.bridgeUrl, entryPoint, JSON.stringify(queryObject));
+NukiBridge.prototype._sendRequest = function _sendRequest(entryPoint, queryObject, requestTimeout, callback /*(err, json)*/) {
+    this.log("Send request to Nuki bridge '%s' on '%s' with '%s'.", this.bridgeUrl, entryPoint, JSON.stringify(queryObject));
     this.runningRequest = true;
     request.get({
         url: this.bridgeUrl+entryPoint,
@@ -176,15 +238,20 @@ NukiBridge.prototype._sendRequest = function _sendRequest(nukiLock, entryPoint, 
         timeout: requestTimeout
     }, (function(err, response, body) {
         var statusCode = response && response.statusCode ? response.statusCode: -1;
-        nukiLock.log("Request to Nuki bridge '%s' finished with status code '%s' and body '%s'.", this.bridgeUrl, statusCode, body, err);
+        this.log("Request to Nuki bridge '%s' finished with status code '%s' and body '%s'.", this.bridgeUrl, statusCode, body, err);
         if (!err && statusCode == 200) {
             var json = JSON.parse(body);
             var success = json.success;
-            if(success == "true" || success == true) {
-                callback(null, json);
+            if(success) {
+                if(success == "true" || success == true) {
+                    callback(null, json);
+                }
+                else {
+                    callback(new Error("Request to Nuki bridge was not succesful."));
+                }
             }
             else {
-                callback(new Error("Request to Nuki bridge was not succesful."));
+                callback(null, json);
             }
         }
         else {  
@@ -197,16 +264,18 @@ NukiBridge.prototype._sendRequest = function _sendRequest(nukiLock, entryPoint, 
 
 NukiBridge.prototype._addToQueue = function _addToQueue(queueEntry) {
     var wasReplaced = false;
-    for (var i = 0; i < this.queue.length; i++) {
-        var alreadyQueuedEntry = this.queue[i];
-        var sameLock = queueEntry.nukiLock.id == alreadyQueuedEntry.nukiLock.id;
-        var bothLockAction = queueEntry.lockAction && alreadyQueuedEntry.lockAction;
-        var bothLockState = !queueEntry.lockAction && !alreadyQueuedEntry.lockAction;
-        var needsReplace = sameLock && (bothLockAction || bothLockState);
-        if(needsReplace) {
-            wasReplaced = true;
-            this.queue[i] = queueEntry;
-            alreadyQueuedEntry.callback(new Error("Request to Nuki bridge was canceled due to newer requests of same type."));
+    if(queueEntry.nukiLock) {
+        for (var i = 0; i < this.queue.length; i++) {
+            var alreadyQueuedEntry = this.queue[i];
+            var sameLock = queueEntry.nukiLock.id == alreadyQueuedEntry.nukiLock.id;
+            var bothLockAction = queueEntry.lockAction && alreadyQueuedEntry.lockAction;
+            var bothLockState = !queueEntry.lockAction && !alreadyQueuedEntry.lockAction;
+            var needsReplace = sameLock && (bothLockAction || bothLockState);
+            if(needsReplace) {
+                wasReplaced = true;
+                this.queue[i] = queueEntry;
+                alreadyQueuedEntry.callback(new Error("Request to Nuki bridge was canceled due to newer requests of same type."));
+            }
         }
     }
     if(!wasReplaced) {
@@ -217,11 +286,19 @@ NukiBridge.prototype._addToQueue = function _addToQueue(queueEntry) {
 NukiBridge.prototype._processNextQueueEntry = function _processNextQueueEntry() {
     if(this.queue.length > 0) {
         var queueEntry = this.queue.shift();
-        if(queueEntry.lockAction) {
-            this.lockAction(queueEntry.nukiLock, queueEntry.lockAction, queueEntry.callback);
+        if(queueEntry.nukiLock) {
+            if(queueEntry.lockAction) {
+                this.lockAction(queueEntry.nukiLock, queueEntry.lockAction, queueEntry.callback);
+            }
+            else {
+                this.lockState(queueEntry.nukiLock, queueEntry.callback);
+            }
         }
-        else {
-            this.lockState(queueEntry.nukiLock, queueEntry.callback);
+        else if (queueEntry.callbackList) {
+            this._getCallbacks(queueEntry.callbackList);
+        }
+        else if (queueEntry.callbackAdd) {
+            this._addCallback();
         }
     }
 };